+++ pr_agent/git_providers/bitbucket_provider.py
@@ -153,6 +153,23 @@ class BitbucketProvider(GitProvider):
         self.diff_files = diff_files
         return diff_files
 
+    def publish_persistent_comment(self, pr_comment: str, initial_text: str, updated_text: str):
+        try:
+            for comment in self.pr.comments():
+                body = comment.raw
+                if initial_text in body:
+                    if updated_text:
+                        pr_comment_updated = pr_comment.replace(initial_text, updated_text)
+                    else:
+                        pr_comment_updated = pr_comment
+                    d = {"content": {"raw": pr_comment_updated}}
+                    response = comment._update_data(comment.put(None, data=d))
+                    return
+        except Exception as e:
+            get_logger().exception(f"Failed to update persistent review, error: {e}")
+            pass
+        self.publish_comment(pr_comment)
+
     def publish_comment(self, pr_comment: str, is_temporary: bool = False):
         comment = self.pr.comment(pr_comment)
         if is_temporary: