+++ pr_agent/git_providers/gitlab_provider.py
@@ -136,6 +136,21 @@ class GitLabProvider(GitProvider):
         except Exception as e:
             get_logger().exception(f"Could not update merge request {self.id_mr} description: {e}")
 
+    def publish_persistent_comment(self, pr_comment: str, initial_text: str, updated_text: str):
+        try:
+            for comment in self.mr.notes.list(get_all=True)[::-1]:
+                if comment.body.startswith(initial_text):
+                    if updated_text:
+                        pr_comment_updated = pr_comment.replace(initial_text, updated_text)
+                    else:
+                        pr_comment_updated = pr_comment
+                    response = self.mr.notes.update(comment.id, {'body': pr_comment_updated})
+                    return
+        except Exception as e:
+            get_logger().exception(f"Failed to update persistent review, error: {e}")
+            pass
+        self.publish_comment(pr_comment)
+
     def publish_comment(self, mr_comment: str, is_temporary: bool = False):
         comment = self.mr.notes.create({'body': mr_comment})
         if is_temporary: